add readmes and cleanup hokulea clients
