+++ bin/client/README.md
@@ -1,0 +1,3 @@
+# `hokulea-client-bin`
+
+This binary contains the client program for executing the Optimism rollup state transition that contains eigenda logics.
