+++ bin/client/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod witness;
-pub mod witgen;
+pub mod clients;
