+++ bin/host/README.md
@@ -1,0 +1,3 @@
+# `hokulea-host-bin`
+
+hokulea-host-bin is a CLI application that runs the pre-image server and client-program from hokulea-client-bin.
