+++ crates/proof/README.md
@@ -1,0 +1,3 @@
+# hokulea-proof
+
+hokulea-proof is a eigenda derivation proof module.
